+++ ios/Runner/Info.plist
@@ -49,11 +49,5 @@
 	</array>
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
-    <key>UIBackgroundModes</key>
-    <array>
-        <string>fetch</string>
-    </array>
-    <key>UIFileSharingEnabled</key>
-    <true/>
 </dict>
 </plist>
